removed deprecated PublishedAtDisabled variable
the published_at input is only disabled if the journal_entry is published, any other calculation is nice, but too complex.
the content managers have to know what they do if they change the published_at input after having published it.
(note: permalink is made with the published_at date)